Desktop: Combine Joplin Cloud and Server OAuth screens resolving PR review
